requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.4
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
idna==3.7
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.0
stripe==10.2.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2